Rewrite Server API

+++ src/main/java/ch/irrigation_system/repository/UnitRepository.java
@@ -3,5 +3,5 @@ package ch.irrigation_system.repository;
 import ch.irrigation_system.model.Unit;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UnitRepository extends CrudRepository<Unit, Integer> {
+public interface UnitRepository extends CrudRepository<Unit, Long> {
 }
